refactor: raceStep을 Race 객체의 책임으로 넘김
