+++ src/controllers/RaceController.js
@@ -1,39 +1,31 @@
-import { ErrorMessages } from '../constants/Messages';
-import Race from '../models/Race';
-import InputView from '../views/InputView';
-import OutputView from '../views/OutputView';
+import Race from '../domain/Race';
+import InputView from '../view/InputView';
+import OutputView from '../view/OutputView';
 
 class RaceController {
   #race;
-  #step;
 
   async startSetup() {
     const carNames = await InputView.readCarNames();
     const raceStep = await InputView.readRaceStep();
 
-    RaceController.#validateRaceStep(raceStep);
-
     this.#race = new Race(carNames);
-    this.#step = raceStep;
-  }
-
-  static #validateRaceStep(raceStep) {
-    if (!Number.isInteger(raceStep)) throw new Error(ErrorMessages.RACE_STEP_NOT_INTEGER);
-
-    if (Number(raceStep) < 0) throw new Error(ErrorMessages.RACE_STEP_NOT_POSITIVE);
+    this.#race.setRaceStep(raceStep);
   }
 
   startRace() {
-    OutputView.printRaceStateTitle();
+    OutputView.printRaceTitle();
 
-    for (let i = 0; i < this.#step; i += 1) {
+    while (!this.#race.isRaceEnd()) {
       this.#race.moveOnce();
-      OutputView.printRaceState(this.#race.getCars());
+
+      const raceStates = this.#race.getRaceStates();
+      OutputView.printRaceState(raceStates);
     }
   }
 
   endRace() {
-    OutputView.printWinners(this.#race.getWinners());
+    OutputView.printWinners(this.#race.findWinnerNames());
   }
 }
